refactor: shared config to server side config
